backend/requirements.txt

pandas
mysql-connector-python
flask
gunicorn